requirements.txt

pandas
openpyxl
python-telegram-bot
xlsxwriter
